requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.2
django-appconf==1.0.3
django-compressor==2.3
django-filter==2.2.0
django-heroku==0.3.1
django-seed==0.1.9
djangorestframework==3.11.0
Faker==3.0.0
gunicorn==20.0.4
idna==2.8
libsass==0.19.3
pi==0.1.2
Pillow==6.2.0
psycopg2-binary
python-dateutil==2.8.1
pytz==2019.3
rcssmin==1.0.6
reportlab==3.5.28
requests==2.22.0
rjsmin==1.1.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.3
Wand==0.5.7
whitenoise==5.0.1
